make length int64 for better interop with System.IO.FileInfo

+++ src/app/FakeLib/Appcast.fs
@@ -42,7 +42,7 @@ type AppcastItem = {
     /// Optional miminal system version for the update
     minimumSystemVersion : string option;
     /// Length of the file in bytes
-    length : uint32;
+    length : int64;
 }
 
 /// Configuration data for the appcast
